Added Exception Handlers

+++ src/main/java/com/stackroute/MovieApp/service/MovieServiceImpl.java
@@ -1,6 +1,8 @@
 package com.stackroute.MovieApp.service;
 
 import com.stackroute.MovieApp.domain.Movie;
+import com.stackroute.MovieApp.exception.MovieAlreadyExistsException;
+import com.stackroute.MovieApp.exception.MovieNotFoundException;
 import com.stackroute.MovieApp.repository.MovieRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,24 +19,37 @@ public class MovieServiceImpl implements MovieService {
     }
 
     @Override
-    public Movie saveMovie(Movie movie) {
-        Movie savedMovie = movieRepository.save(movie);
-        return savedMovie;
+    public Movie saveMovie(Movie movie) throws MovieAlreadyExistsException {
+           if(movieRepository.existsById(movie.getId())){
+               throw new MovieAlreadyExistsException("Movie Already Exists");
+           }
+            Movie savedMovie = movieRepository.save(movie);
+            return savedMovie;
+
     }
 
     @Override
-    public List<Movie> getAllMovies() {
+    public List<Movie> getAllMovies() throws MovieNotFoundException {
+        if(movieRepository.count()==0){
+            throw new MovieNotFoundException("Movies not found");
+        }
         return movieRepository.findAll();
     }
 
     @Override
-    public boolean deleteMovie(int id) {
+    public boolean deleteMovie(int id) throws MovieNotFoundException{
+        if(movieRepository.existsById(id)){
+            throw new MovieNotFoundException("Movie to be deleted not found");
+        }
         movieRepository.deleteById(id);
         return true;
     }
 
     @Override
-    public Movie updateMovie(Movie movie) {
+    public Movie updateMovie(Movie movie) throws MovieNotFoundException{
+        if(movieRepository.existsById(movie.getId())){
+            throw new MovieNotFoundException("Movie to be updated not found");
+        }
         Movie updatedMovie=movieRepository.getOne(movie.getId());
         updatedMovie.setMovieName(movie.getMovieName());
         updatedMovie.setPlot(movie.getPlot());
